pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[project]
name = "crypto-multi-trader"
version = "0.1.0"
description = "멀티 계정 비트코인 자동매매 봇"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "bcrypt>=4.0.0",
    "itsdangerous>=2.1.0",
    "starlette-csrf>=3.0.0",
    "python-binance>=1.0.20",
    "aiolimiter>=1.1.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "cryptography>=42.0.0",
    "pandas>=2.0.0",
    "prometheus-client>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "pytest-cov>=5.0.0",
    "coverage[toml]>=7.4.0",
    "ruff>=0.3.0",
    "pre-commit>=3.6.0",
    "factory-boy>=3.3.0",
    "time-machine>=2.13.0",
    "sentry-sdk[fastapi]>=2.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "unit: pure logic tests, no I/O",
    "integration: requires test database",
    "slow: tests taking >5s",
    "exchange: tests involving exchange client",
]
addopts = [
    "--strict-markers",
    "--tb=short",
    "-q",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning:binance.*",
    "ignore::DeprecationWarning:websockets.*",
    "ignore::pydantic.warnings.PydanticDeprecatedSince20",
    "ignore::DeprecationWarning:alembic.*",
]

[tool.coverage.run]
source = ["app"]
omit = [
    "app/dashboard/*",
    "app/models/base.py",
    "alembic/*",
]
branch = true

[tool.coverage.report]
precision = 1
show_missing = true
skip_covered = true
fail_under = 0

[tool.coverage.html]
directory = "htmlcov"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument (FastAPI Depends pattern)
    "B904",   # raise from within except (existing code style)
]
